Change IntColumn to RealColumn for financial fields
- Change basicSalary in Employees from IntColumn to RealColumn
- Change amount in Expenses, CashTransactions, and Payments from IntColumn to RealColumn
- Revert all .toInt() conversions back to .toDouble()
- This matches the generated code in local_db.g.dart which uses double for these fields

+++ mobile/lib/services/salary_entitlement_service.dart
@@ -83,21 +83,21 @@ class SalaryEntitlementService {
     for (final expense in expenses) {
       final type = expense.expenseType.trim();
       if (type == 'سحب راتب' || type == 'سلفة' || type == 'رواتب') {
-        totalWithdrawals += expense.amount.toDouble();
+        totalWithdrawals += expense.amount;
         transactions.add(
           SalaryTransaction(
             type: 'سحب',
-            amount: expense.amount.toDouble(),
+            amount: expense.amount,
             date: expense.date,
             note: expense.description,
           ),
         );
       } else if (type == 'خصم راتب' || type == 'خصم' || type == 'غياب') {
-        totalDeductions += expense.amount.toDouble();
+        totalDeductions += expense.amount;
         transactions.add(
           SalaryTransaction(
             type: 'خصم',
-            amount: expense.amount.toDouble(),
+            amount: expense.amount,
             date: expense.date,
             note: expense.description,
           ),
@@ -113,7 +113,7 @@ class SalaryEntitlementService {
       employee: employee,
       hireDate: hireDate,
       totalMonthsWorked: totalMonthsWorked,
-      basicSalary: employee.basicSalary.toDouble(),
+      basicSalary: employee.basicSalary,
       totalEntitlement: totalEntitlement,
       totalWithdrawals: totalWithdrawals,
       totalDeductions: totalDeductions,
